/install: extract QuickInstall, OS-tab everything, add yarn/deno, fix hydration
Hydration error (1 + 2 in the user's list):
- `release.publishedAt` was rendered with `toLocaleDateString(undefined,
  …)`, which picks up the SSR runtime's locale (en-US, often UTC) on
  the server vs. the browser's locale on the client. The mismatched
  text node tripped React 19's hydration check. Pinned the locale to
  `"en-US"` and `timeZone: "UTC"` so SSR and the browser produce
  identical output. Same fix removes the secondary "script tag while
  rendering" warning since both originate from the same hydration
  desync.

Banner spacing (3):
- Tightened `.install-hero-meta` gap (1rem → 0.5rem) and dropped the
  legacy `.install-meta { margin: 0.5rem 0 0 }` and `.prestable-banner
  { margin: 0.5rem 0 2rem }` so the two cards sit close together
  inside the hero's right column rather than floating apart.

Quick install reuse (4):
- Extracted `<QuickInstall>` into `components/quick-install.tsx`.
- Extracted shared `<CommandTabs>`, `OS_TABS`, `PM_TABS`, and
  `detectOs` into `components/command-tabs.tsx` so both the
  homepage and the install page import from one source.
- Replaced the homepage's inline `CurlInstall` with `<QuickInstall>`,
  followed by a small meta row holding the latest-release pill and
  the "Other install options →" link to `/install`.

OS-tabbed long install methods (5 + 7 + 8):
- Consolidated Homebrew / apt / Scoop into a single "Native package
  manager" method with OS tabs.
- "Pre-built binaries" now OS-tabs and downloads the **full
  toolchain** (`GocciaScriptLoader` + `GocciaTestRunner` +
  `GocciaREPL`). macOS/Linux use a `for` loop and `chmod +x` →
  `/usr/local/bin/`; Windows uses a PowerShell `foreach` over
  `Invoke-WebRequest` writing to `$env:USERPROFILE\bin`.
- "Build from source" OS-tabs the final command (Unix
  `./build/GocciaScriptLoader --help` vs. Windows
  `.\\build\\GocciaScriptLoader.exe --help`); the build line itself
  (`./build.pas loader testrunner repl`) is identical.
- All four OS-tabbed methods share the `goccia.install.os` storage
  key so picking macOS in one updates the others — same as the PM
  picker's `goccia.install.pm` key.

yarn + deno (6):
- Added `YarnIcon` (yarn ball-of-yarn) and `DenoIcon` (Deno dinosaur
  silhouette) using simple-icons paths (CC0).
- Extended `PM_TABS` to npm / Bun / pnpm / Yarn / Deno.
- Run-without-installing: yarn → `yarn dlx gocciascript script.js`,
  deno → `deno run -A npm:gocciascript script.js`.
- JS package managers: yarn → `yarn global add gocciascript`,
  deno → `deno install -gA -n gocciascript npm:gocciascript`.

`$PATH` rendering + FreePascal link (9 + 10):
- Every `$PATH` mention is now `<code>$PATH</code>` (the install hero
  intro already had this; "Build from source" now does too).
- "FreePascal compiler" in the Build-from-source description is now
  a link to https://www.freepascal.org with `target="_blank" rel="noopener noreferrer"`.

Verification: lint + tsc + build clean, 60 tests pass.
