+++ website/src/app/globals.css
@@ -689,7 +689,7 @@ p {
      max-width keeps it readable on very wide single-column layouts
      (e.g. mobile portrait → desktop transition). */
   max-width: 70ch;
-  margin: 0.5rem 0 2rem;
+  margin: 0;
   padding: 1rem 1.25rem;
   background: color-mix(in oklab, var(--amber) 12%, var(--paper-2));
   border: 1px solid color-mix(in oklab, var(--amber) 40%, var(--rule-soft));
@@ -740,7 +740,7 @@ p {
 .install-hero-meta {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 0.5rem;
 }
 
 /* The featured first method (Quick install — one-liner) gets a soft
@@ -754,6 +754,21 @@ p {
 }
 .install-method.install-method-featured h2 {
   margin-top: 0;
+}
+
+/* The Quick install component is reused on the homepage; this row
+   places the version pill + "other install options" link beneath the
+   tabbed installer at hero scale. */
+.hero-quick-install {
+  display: flex;
+  flex-direction: column;
+  gap: 0.85rem;
+}
+.hero-install-meta {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 0.85rem;
 }
 /* Generic tabbed command picker — used for both the OS picker
    (Quick install card) and the package-manager picker (npm / Bun /
@@ -837,7 +852,7 @@ p {
   display: flex;
   flex-direction: column;
   gap: 0.6rem;
-  margin: 0.5rem 0 0;
+  margin: 0;
 }
 .install-latest {
   margin: 0;
